+++ lua/plugins/neo-tree.lua
@@ -1,0 +1,37 @@
+-- Neo-tree Overrides
+
+if true then return {} end -- WARN: REMOVE THIS LINE TO ACTIVATE THIS FILE
+
+---@type LazySpec
+return {
+  "nvim-neo-tree/neo-tree.nvim",
+  config = function()
+    require("neo-tree").setup {
+      filesystem = {
+        filtered_items = {
+          visible = true, -- show hidden files in alternate style
+          hide_dotfiles = true,
+          hide_gitignored = true,
+          hide_hidden = true, -- only works on Windows for hidden files/directories
+          hide_by_name = {
+            --"node_modules"
+          },
+          hide_by_pattern = { -- uses glob style patterns
+            --"*.meta",
+            --"*/src/*/tsconfig.json",
+          },
+          always_show = { -- remains visible even if other settings would normally hide it
+            --".gitignored",
+          },
+          never_show = { -- remains hidden even if visible is toggled to true, this overrides always_show
+            --".DS_Store",
+            --"thumbs.db"
+          },
+          never_show_by_pattern = { -- uses glob style patterns
+            --".null-ls_*",
+          },
+        },
+      },
+    }
+  end,
+}